project/api/requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
pydantic==2.8.2
motor==3.5.1
python-dotenv==1.0.1
pytz
aioredis
redis